cli refactoring and exporting high-level api
