+++ utensor_cgen/frontend/base.py
@@ -17,5 +17,5 @@ class Parser(object):
 
   @classmethod
   @abstractmethod
-  def parse(cls, fname, outupt_nodes):
+  def parse(cls, fname, outupt_nodes, model_name=None):
       raise RuntimeError('abstract parse method involded')
